RPRMAX-391: Uber material, Slow performance with high polygon count (#130)
* RPRMAX-391: Uber material, Slow performance with high polygon count

Added handling in FireRenderMtlBase base class to overrive method Mtl::Requirements().  The override method now calls the superclass method and caches the result.  The cache is meant to be discarded whenever the material or its subtextures/submaterials change.  Changes are currently detected as any call to NotifyRefChanged().

This allows Requirements() to be called millions of times with near-zero overhead.  The default implementation of Mtl::Requirements() calls GetSubTexmap(), which FireRender materials implement by performing an STL map lookup via this->TEXMAP_MAPPING.find(i).  Previously, without caching and due to being called millions of times, the lookups incurred significant performance impact.

REMAINING ISSUES:

Hypothetically, there may be cases where a change occurs without triggering FireRenderMtlBase::NotifyRefChanged(), for example if a subclass implements NotifyRefChanged() and fails to call the base class method.  If such cases also cause the material requirements to change, then FireRenderMtlBase::Requirements() would hold an invalid cache report an incorrect result.  For this reason, FireRenderMtlBase::NotifyRefChanged() is now marked as final, but please be alert to other possible special cases in the future.

* Changed variables to strict camel case naming

Per request from Denis Kolliakov.  As a note, I personally prefer underscores whenever a single identifier name conveys two disjoint concepts.  In this case the _requirements postfix was meant to indicate the context of the Requirements function.  The underscore here serves as a separator, analagous to a scope operator.  For example, if ten functions each required copies of the changedCounter variable for caching, this would visually emphasize that ten variants of the changedCount exist.  This is now switched to the word "For" instead of underscore for similar effect.