RPRMAX-350: Ambient occlusion node spec and implementation (#137)
Added support for the Ambient Occlusion shader, available as a texture map named RPR Ambient Occlusion, in the Radeon ProRender category, from the Material/Map Browser.  Exposed six UI parameters for radius, side which may be front or back, unoccluded color and texture, and occluded color and texture.  During export, the texture is translated to a BlendValue shader with an AmbientOcclusion shader attached as the weight input value, and with unoccluded and occluded textures attached as the color input values.

REMAINING ISSUES:

* No UI exposure added for the system parameter Ambient Occlusion Ray Length, which is internally named RPR_CONTEXT_AO_RAY_LENGTH.  This could be exposed though the render settings or elsewhere.  To be implemented pending further consideration.

* No support added for XML import or export of the Ambient Occlusion texture map.  XML export does not appear to be working at present.  Thus testing and validation were not possible.   To be implemented pending further consideration.