RPRMAX-406: ActiveShade, support for NodeEventSystem (#139)
* RPRMAX-406: ActiveShade, support for NodeEventSystem

Work in progress, but currently disabled.  Added support for the NodeEventSystem in Radeon ProRender with ActiveShade.  Replaced uses of standard 3ds Max reference system with uses of the NodeEventSystem.  In testing, this appears stable with no issues found.  However, no significant performance gain was measured either.  As the performance gain requires further evaluation, and due to current release schedule, the changed are disabled internally, pending further consideration.

TECHNICAL DETAILS:
* Added class SynchronizerNodeEventCallback as superclass of Synchronizer.  This class and its methods are analogous to the handling in NotifyRefChanged(), receiving notifications of scene changes, and posting commands to rebuild corresponding RPR items.  In the future, if NodeEventSystem handling is enabled, most handling in NotifyRefChanged() could be removed.

* Added macro FOR_EACH_NODE.  This safely handles objects which have been deleted, a scenario which may occur because notifications are delayed from scene events.  Deleted nodes maintain valid node keys (AnimHandles), which are never reused by more recently created nodes.  The function GetNodeByKey() detects a deleted node, and responds with NULL.

* Added callback key,  class Synchronizer member mNodeEventCallbackKey.  This holds the registration key from the NodeEventSystem.  It is needed to unregister later.  Initialized to -1 since valid keys are non-negative.

* Fixed build error (illegal reference to non-static member)

Also updated handling for mNodeEventCallbackKey.  Replaced checks for "greater than or equal to zero" with "not equal to -1".  The variable is an unsigned DWORD, therefore it will always be greater than or equal to zero.  However, keys are assigned sequentially.  Since -1 is interpreted as the maximum DWORD value, this is a reasonable initialization value, because callback keys should never be nearly this large, as that would require more than four billion callback registrations in one session.