RPRMAX-466:  Support OSL shaders in 3ds Max
Work in progress.  Added initial support for the Open Shading Language (OSL) shader, available previously in 3ds max, and exposed in the Radeon ProRender SDK as the User Texture.

Currently requires that a compiled shader .INL file is provided.  Does not yet invoke the shader compiler.  During export, the .OSL source filename of the texture is matched to a compiled .INL file, which must be named "NAME_OutputN.inl", where NAME is source file, and N is the output selected.  For example, if a scene uses OSL shader "Rivets.osl", and the color channel (first output) is connected to an input to another texture, the exporter regards this as "Rivets" shader output "0", and then searches for a file named "Rivets_Output0.inl" in the root install folder of 3ds max.  Also, in this case, the 3ds max texture graph contains a Multi Output To Texmap texture holding the OSL shader texture as a subtexture.  Since textures in 3ds max have only one output, this is required to allow a particular OSL output channel to be selected for the output.  The exporter queries the Multi Output to Texture to determine the output index number N, as above.

TECHNICAL DETAILS:

* Each OSL shader file is considered a unique texture type in RPR.  Currently this is limited to four unique textures, RPR_MATERIAL_NODE_USER_TEXTURE_0 through RPR_MATERIAL_NODE_USER_TEXTURE_3.  Therefore, only four OSL shader files may be used in a scene, or even fewer, where multiple output channels are required from a single shader.  During export, a resource number is assigned from 0-3 for each .INL file used.  As above, if multiple outputs of an .OSL shader are used, different resources (.INL files) are required for each.

* During export, a mapping is created from parameter names to shader input indices.  The OSL Shader texture has two parameter blocks; the first with the shader filename, the second with shader parameters.  The latter has typically two entries for each shader input, one with the static value and the second with the connected texture, if any.  These are provided in original order.  The static value may not be available for all parameters, but the connected texture parameter is always provided, and named "PARAM_map" where PARAM is the original shader parameter.  The correct shader input order is determined by iterating through the parameters, then associating one input index, in order listed, to each parameter of texture type.

* Compiled shader code is passed to RPR via rprContextSetParameterString(handle,"textureN",code) where N is the resource number, from 0-3 as above, and code is a string holding the .INL file contents.