RPRMAX-528: Hardware rollout resetting (#185)
Done.  Stores Hardware rollout preferences in the attribute settings file, \en-us\plugcfg\Radeon ProRender\FireRenderSettings.ini.  Settings follow the same conventions as in Maya:
- DevicesName, comma-separated list of GPU device names as shown in the Hardware rollout
- DevicesSelected, comma-separated list of boolean values (0 or 1) indicating if the corresponding device is selected
- OverrideCPUThreadCount, value of the CPU thread count override checkbox (0 or 1)
- CPUThreadCount, value of the CPU thread count spinner
When settings are loaded, the GPU names from the preferences are compared with GPU names of the current hardware, and matched up.  So, if the hardware is changed, the toggle settings will be preserved as long as GPU names remain the same.  If none of the previously toggled GPUs are present anymore, then only CPU will be toggled in the UI.  If multiple identical GPUs exist but only some were toggled, the preferences record how many, but not which were toggled, so when next loaded, the matching checkboxes appearing first in the UI will be toggled first.