RPRMAX-570: Production render gives Black render in interior scene (#199)
Fixed.  Scene was saved with minimum samples parameter value of zero.  That setting is lower than should be possible, and yields a black render.  This is fixed by enforcing a minimum of 2.  Per recommendation from Takahiro Harada, the minimum should be no less than 8; although a minimum of 2 is enough to produce a render image, it could be increased in the future.